refactor(events): extract catalog-agnostic event-stream core (TS + Go)
Pull the transport/registry out of the tutti-coupled packages into two neutral,
reusable cores so other products (e.g. tsh group chat) can share them.

- @tutti-os/event-stream-core (TS, publishable): connection lifecycle, heartbeat,
  backoff reconnect, frame codec, subscription registry. Generic over the
  client/server event + scope types, parameterized by an injected
  EventStreamProtocol. client-tuttid-ts keeps its public
  createTuttidEventStreamClient surface and becomes a thin workspace
  (scope.workspaceId) binding on top.
- streamgo (Go module packages/events/stream-go): generic Service[S comparable]
  pub/sub registry + scope routing + session fan-out. tuttid eventstream becomes
  a thin binding via type aliases (daemon_events.go unchanged); scope axis is now
  generic and the catalog dependency is reduced to a minimal TopicVersion contract.

Also externalize @tutti-os/event-stream-core in the Electron build (it joins the
client-tuttid-ts source graph consumed by the main process).

No behavior change: eventStreamClient.test.ts (13/13) and tuttid eventstream/api
tests stay green; gofmt + go vet clean. The pnpm-lock.yaml update for the new
workspace dependency is intentionally left to the canonical CI environment (a
local non-Linux install prunes cross-platform optional deps and reflows quoting).
