+++ packages/clients/tuttid-ts/package.json
@@ -11,7 +11,8 @@
     ".": "./src/index.ts"
   },
   "dependencies": {
-    "@tutti-os/event-protocol": "workspace:*"
+    "@tutti-os/event-protocol": "workspace:*",
+    "@tutti-os/event-stream-core": "workspace:*"
   },
   "devDependencies": {
     "@types/node": "^24.0.1",
