+++ packages/events/stream-core/src/index.ts
@@ -1,0 +1,14 @@
+export {
+  createEventStreamClient,
+  type CreateEventStreamClientInput,
+  type EventStreamClient,
+  type EventStreamClientFrame,
+  type EventStreamConnectionState,
+  type EventStreamHeartbeatConfig,
+  type EventStreamProtocol,
+  type EventStreamReconnectConfig,
+  type EventStreamServerEvent,
+  type EventStreamServerFrame,
+  type EventStreamSocket,
+  type EventStreamSocketFactory
+} from "./eventStreamClient.ts";
