+++ packages/events/stream-core/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../../packages/configs/tsconfig/base.json",
+  "compilerOptions": {
+    "allowImportingTsExtensions": true
+  },
+  "include": ["src/**/*"]
+}
