+++ packages/events/stream-core/tsup.config.ts
@@ -1,0 +1,11 @@
+import { defineConfig } from "tsup";
+
+export default defineConfig({
+  clean: true,
+  dts: true,
+  entry: {
+    index: "src/index.ts"
+  },
+  format: ["esm"],
+  sourcemap: true
+});
