+++ packages/events/stream-go/go.mod
@@ -1,0 +1,3 @@
+module github.com/tutti-os/tutti/packages/events/stream-go
+
+go 1.24.3
