+++ services/tuttid/go.mod
@@ -11,6 +11,7 @@ require (
 	github.com/oapi-codegen/runtime v1.4.1
 	github.com/tutti-os/tutti/packages/agentactivity/daemon v0.0.0
 	github.com/tutti-os/tutti/packages/appcli/core v0.0.0
+	github.com/tutti-os/tutti/packages/events/stream-go v0.0.0
 	github.com/tutti-os/tutti/packages/workbench/service v0.0.0
 	github.com/tutti-os/tutti/packages/workspace/files v0.0.0
 	github.com/tutti-os/tutti/packages/workspace/issues v0.0.0
@@ -18,6 +19,8 @@ require (
 	golang.org/x/sys v0.41.0
 	modernc.org/sqlite v1.45.0
 )
+
+replace github.com/tutti-os/tutti/packages/events/stream-go => ../../packages/events/stream-go
 
 replace github.com/tutti-os/tutti/packages/workbench/service => ../../packages/workbench/service
